show float num ok

+++ HARDWARE/ADC/adc.h
@@ -18,7 +18,7 @@ extern u32 min_data;
  							   
 void  ADC_DMA_Init(void);
 void Get_Value(void);
-u16 ADC_Get_Vpp(void);
+float ADC_Get_Vpp(void);
 
 #endif
